Break down large test file
